Standardize to unsigned int

+++ main.cu
@@ -13,13 +13,13 @@ using namespace std;
 
 #define NUM_THREADS 512 // El número mínimo de threads es 32 (por el tamaño de warp) y el maximo 1024
 
-void initAllProductsBuyOptions(int *all_products_buy_options);
+void initAllProductsBuyOptions(unsigned int *all_products_buy_options);
 
-void printAllProductsAllBuyOptions(int *all_products_buy_options);
-void getBestBuyOptions(int *all_products_buy_options, int *best_buy_options);
-void printBestBuyOptions(int *best_buy_options);
+void printAllProductsAllBuyOptions(unsigned int *all_products_buy_options);
+void getBestBuyOptions(unsigned int *all_products_buy_options, unsigned int *best_buy_options);
+void printBestBuyOptions(unsigned int *best_buy_options);
 
-bool areResultsValid(int *all_products_buy_options, int *best_buy_options);
+bool areResultsValid(unsigned int *all_products_buy_options, unsigned int *best_buy_options);
 
 // ToDo: Cada thread ejecuta el kernel.
 // Identificar el thread en el que estamos, en base a esto, calculamos si nos toca trabajar y, en caso afirmativo
@@ -29,7 +29,7 @@ bool areResultsValid(int *all_products_buy_options, int *best_buy_options);
 
 __global__ void KernelKnapsack(unsigned int *total_buy_options, unsigned int *best_buy_options)
 {
-    __shared__ float tmp_best_buy_options[1024];
+    __shared__ unsigned int tmp_best_buy_options[1024];
     unsigned int stride;
 
     unsigned int thread_id = threadIdx.x;
@@ -64,13 +64,13 @@ __global__ void KernelKnapsack(unsigned int *total_buy_options, unsigned int *be
 int main(int argc, char** argv)
 {
     // Buy options in host and device
-    int *host_all_products_buy_options = (int *) malloc( NUM_PRODUCTS * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION * sizeof(int) );
-    int *best_buy_options = (int *) malloc( NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION * sizeof(int) );
+    unsigned int *host_all_products_buy_options = (unsigned int *) malloc( NUM_PRODUCTS * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION * sizeof(unsigned int) );
+    unsigned int *best_buy_options = (unsigned int *) malloc( NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION * sizeof(unsigned int) );
     unsigned int *device_all_products_buy_options;
     unsigned int *device_best_buy_options;
 
     // Metadata
-    unsigned int buy_option_size = ELEMENTS_PER_BUY_OPTION * sizeof(int);
+    unsigned int buy_option_size = ELEMENTS_PER_BUY_OPTION * sizeof(unsigned int);
     unsigned int num_total_buy_options = NUM_PRODUCTS * NUM_BUY_OPTIONS;
     unsigned int total_buy_options_size = num_total_buy_options * buy_option_size;
     unsigned int best_buy_options_size = NUM_PRODUCTS * buy_option_size;
@@ -86,8 +86,8 @@ int main(int argc, char** argv)
     cudaEventCreate(&stop);
 
     // Obtener Memoria en el device
-    cudaMalloc( (float**) &device_all_products_buy_options, total_buy_options_size );
-    cudaMalloc( (float**) &device_best_buy_options, best_buy_options_size );
+    cudaMalloc( (unsigned int**) &device_all_products_buy_options, total_buy_options_size );
+    cudaMalloc( (unsigned int**) &device_best_buy_options, best_buy_options_size );
 
     // Copiar datos desde el host en el device
     cudaMemcpy(device_all_products_buy_options, host_all_products_buy_options, total_buy_options_size, cudaMemcpyHostToDevice);
@@ -114,7 +114,7 @@ int main(int argc, char** argv)
     printf("Number of Threads: %d\n", NUM_THREADS);
     printf("Number of blocks (products): %d\n", NUM_PRODUCTS);
     printf("Total time %4.6f milseg\n", elapsed_time);
-    printf("Bandwidth %4.3f GB/s\n", (num_total_buy_options * sizeof(int)) / (1000000 * elapsed_time));
+    printf("Bandwidth %4.3f GB/s\n", (num_total_buy_options * sizeof(unsigned int)) / (1000000 * elapsed_time));
 
     cudaEventDestroy(start);
     cudaEventDestroy(stop);
@@ -135,7 +135,7 @@ int main(int argc, char** argv)
     }
 }
 
-void initAllProductsBuyOptions(int *all_products_buy_options)
+void initAllProductsBuyOptions(unsigned int *all_products_buy_options)
 {
     // START Vector all the buy options for the product 1
 
@@ -162,23 +162,23 @@ void initAllProductsBuyOptions(int *all_products_buy_options)
     // END Vector all the buy options for the product 2
 }
 
-bool areResultsValid(int *all_products_buy_options, int *best_buy_options)
+bool areResultsValid(unsigned int *all_products_buy_options, unsigned int *best_buy_options)
 {
-   int *tmp_best_buy_options;
+   unsigned int *tmp_best_buy_options;
 
    getBestBuyOptions(all_products_buy_options, tmp_best_buy_options);
 
-   for (int product_iteration = 0; product_iteration < NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION; product_iteration += ELEMENTS_PER_BUY_OPTION)
+   for (unsigned int product_iteration = 0; product_iteration < NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION; product_iteration += ELEMENTS_PER_BUY_OPTION)
    {
-       int current_product_position = product_iteration * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION;
-       int current_product_store_position = current_product_position + STORE_ID_OFFSET;
-       int current_product_price_position = current_product_position + PRICE_OFFSET;
+       unsigned int current_product_position = product_iteration * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION;
+       unsigned int current_product_store_position = current_product_position + STORE_ID_OFFSET;
+       unsigned int current_product_price_position = current_product_position + PRICE_OFFSET;
 
-       int best_store = best_buy_options[current_product_store_position];
-       int best_price = best_buy_options[current_product_price_position];
+       unsigned int best_store = best_buy_options[current_product_store_position];
+       unsigned int best_price = best_buy_options[current_product_price_position];
 
-       int tmp_best_store = tmp_best_buy_options[current_product_store_position];
-       int tmp_best_price = tmp_best_buy_options[current_product_price_position];
+       unsigned int tmp_best_store = tmp_best_buy_options[current_product_store_position];
+       unsigned int tmp_best_price = tmp_best_buy_options[current_product_price_position];
 
        if (best_store != tmp_best_store || best_price != tmp_best_price)
        {
@@ -189,12 +189,12 @@ bool areResultsValid(int *all_products_buy_options, int *best_buy_options)
    return true;
 }
 
-void printAllProductsAllBuyOptions(int *all_products_buy_options)
+void printAllProductsAllBuyOptions(unsigned int *all_products_buy_options)
 {
     cout << "All products buy options:" << endl;
-    for (int i = 0; i < NUM_PRODUCTS; ++i){
+    for (unsigned int i = 0; i < NUM_PRODUCTS; ++i){
         cout << endl << "\tproduct_id: " << i << endl;
-        for (int j = 0; j < NUM_BUY_OPTIONS*2; j += 2){
+        for (unsigned int j = 0; j < NUM_BUY_OPTIONS*2; j += 2){
             cout << "Buy option:" << endl;
             cout << "\tstore_id: " << all_products_buy_options[i*NUM_BUY_OPTIONS*2+j] << endl;
             cout << "\tprice: " << all_products_buy_options[i*NUM_BUY_OPTIONS*2+j+1] << endl;
@@ -202,18 +202,18 @@ void printAllProductsAllBuyOptions(int *all_products_buy_options)
     }
 }
 
-void getBestBuyOptions(int *all_products_buy_options, int *best_buy_options)
+void getBestBuyOptions(unsigned int *all_products_buy_options, unsigned int *best_buy_options)
 {
-    for(int product_iteration = 0; product_iteration < NUM_PRODUCTS; ++product_iteration)
+    for(unsigned int product_iteration = 0; product_iteration < NUM_PRODUCTS; ++product_iteration)
     {
-        int current_product_position = product_iteration * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION;
-        int current_product_store_position = current_product_position + STORE_ID_OFFSET;
-        int current_product_price_position = current_product_position + PRICE_OFFSET;
+        unsigned int current_product_position = product_iteration * NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION;
+        unsigned int current_product_store_position = current_product_position + STORE_ID_OFFSET;
+        unsigned int current_product_price_position = current_product_position + PRICE_OFFSET;
 
-        int best_store = all_products_buy_options[current_product_store_position];
-        int best_price = all_products_buy_options[current_product_price_position];
+        unsigned int best_store = all_products_buy_options[current_product_store_position];
+        unsigned int best_price = all_products_buy_options[current_product_price_position];
 
-        for(int product_to_compare = ELEMENTS_PER_BUY_OPTION; product_to_compare < NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION; product_to_compare += ELEMENTS_PER_BUY_OPTION)
+        for(unsigned int product_to_compare = ELEMENTS_PER_BUY_OPTION; product_to_compare < NUM_BUY_OPTIONS * ELEMENTS_PER_BUY_OPTION; product_to_compare += ELEMENTS_PER_BUY_OPTION)
         {
             if (all_products_buy_options[current_product_position + product_to_compare + PRICE_OFFSET] < best_price)
             {
@@ -226,10 +226,10 @@ void getBestBuyOptions(int *all_products_buy_options, int *best_buy_options)
     }
 }
 
-void printBestBuyOptions(int *best_buy_options)
+void printBestBuyOptions(unsigned int *best_buy_options)
 {
     cout << endl <<"Best products buy options:" << endl;
-    for (int i = 0; i < NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION; i += ELEMENTS_PER_BUY_OPTION)
+    for (unsigned int i = 0; i < NUM_PRODUCTS * ELEMENTS_PER_BUY_OPTION; i += ELEMENTS_PER_BUY_OPTION)
     {
         cout << "Best buy option for product_id: " << i / ELEMENTS_PER_BUY_OPTION << endl;
         cout << "\tstore_id: " << best_buy_options[i + STORE_ID_OFFSET] << endl;
